+++ src/ts/UIComponentConfigOverrides.ts
@@ -2,6 +2,7 @@ import type { UIVariantIdentifier } from './UIManager';
 import type { AdControlBarConfig } from './components/ads/AdControlBar';
 import type { AdCounterLabelConfig } from './components/ads/AdCounterLabel';
 import type { AdSkipButtonConfig } from './components/ads/AdSkipButton';
+import type { OverlayAdStatusOverlayConfig } from './components/ads/OverlayAdStatusOverlay';
 import type { ButtonConfig } from './components/buttons/Button';
 import type { CloseButtonConfig } from './components/buttons/CloseButton';
 import type { PlaybackToggleButtonConfig } from './components/buttons/PlaybackToggleButton';
@@ -183,6 +184,10 @@ export interface UIComponentConfigMap {
    * @category Components
    */
   AdStatusOverlay?: Partial<ContainerConfig>;
+  /**
+   * @category Components
+   */
+  OverlayAdStatusOverlay?: Partial<OverlayAdStatusOverlayConfig>;
   /**
    * @category Components
    */