+++ src/ts/UIFactory.ts
@@ -35,6 +35,7 @@ import { Watermark } from './components/Watermark';
 import { ErrorMessageOverlay } from './components/overlays/ErrorMessageOverlay';
 import { AdClickOverlay } from './components/ads/AdClickOverlay';
 import { AdControlBar } from './components/ads/AdControlBar';
+import { OverlayAdStatusOverlay } from './components/ads/OverlayAdStatusOverlay';
 import { MetadataLabel, MetadataLabelContent } from './components/labels/MetadataLabel';
 import { PlayerUtils } from './utils/PlayerUtils';
 import { CastUIContainer } from './components/CastUIContainer';
@@ -334,6 +335,7 @@ export namespace UIFactory {
           subtitleOverlay,
           new BufferingOverlay(),
           new PlaybackToggleOverlay(),
+          new OverlayAdStatusOverlay(),
           new CastStatusOverlay(),
           controlBar,
           new TitleBar(),
@@ -449,6 +451,7 @@ export namespace UIFactory {
           new CastStatusOverlay(),
           // Use the touch overlay on mobile devices and the regular playback toggle overlay on desktop browsers
           BrowserUtils.isMobile ? new TouchControlOverlay() : new PlaybackToggleOverlay(),
+          new OverlayAdStatusOverlay(),
           new RecommendationOverlay(),
           controlBar,
           new TitleBar({
@@ -646,6 +649,7 @@ export namespace UIFactory {
           subtitleOverlay,
           new BufferingOverlay(),
           playbackToggleOverlay,
+          new OverlayAdStatusOverlay(),
           controlBar,
           titleBar,
           settingsPanel,