+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- `OverlayAdStatusOverlay` to show an ad badge for overlay ads and expose a delayed dismiss button.
+
 ## [4.17.0] - 2026-07-02
 
 ### Added